tests: pin the pending-create contract; delete what contradicted or duplicated it
The tree gains the tests that pin the contract: a parked create
installs at the occupant's retire (engine, EDT, event and data-block
forms; a label reused around a ring, on every kind, with an early lane
and a rotating home); a rendezvous loser parks and is destroyed with
the generation that admitted it; a reservation has one winner; the
destroy roster tells every rank once, size-zero readers and writers
included; a destroyed never-run EDT decrements its scope once; the
finish-scope creator token is released after the join; a first-touch
stub is claimed by the create that follows; a remote finish scope is
refused in Debug builds.

It loses one hundred and twenty-one: programs whose shape the contract
now leaves undefined (a double create with no destroy between, a
create that expected a report, a reader unordered against its
creator's write), duplicates of an oracle that another test already
carries, and tests that asserted nothing the exit status could see.
Every survivor reports by exit status: the printed-verdict regexes are
gone, and a test's main merges the EDT-side failure with what the
runtime saw of the ranks it spawned.

(cherry picked from commit b51ae62ea8f6dbd8760479005d90f0a68a3d083c)

+++ tests/ocr/edt_value.c
@@ -5,7 +5,7 @@ static void check(uint32_t pc, const uint64_t *pv, uint32_t dc,
   if (pc != dc)
     arts_abort(1);
   for (uint32_t i = 0; i < dc; ++i) {
-    if (dv[i].guid != pv[i] || dv[i].ptr != NULL ||
+    if ((uint64_t)dv[i].guid != pv[i] || dv[i].ptr != NULL ||
         dv[i].mode != DB_MODE_NULL) {
       arts_printf("FAIL edt_value: slot %u changed value, mode, or pointer\n", i);
       arts_abort(1);
